+++ extensions/vscode/src/commands/index.ts
@@ -1,4 +1,5 @@
 export * from "./add-brick";
 export * from "./init";
 export * from "./make-brick";
+export * from "./new-brick";
 export * from "./remove-brick";