+++ packages/mason_cli/lib/src/commands/new.dart
@@ -136,6 +136,10 @@ class _BrickGenerator extends MasonGenerator {
 name: $name
 description: $description
 
+# The following defines the brick repository url.
+# Uncomment and update the following line before publishing the brick.
+# repository: https://github.com/my_org/my_repo
+
 # The following defines the version and build number for your brick.
 # A version number is three numbers separated by dots, like 1.2.34
 # followed by an optional build number (separated by a +).
@@ -180,11 +184,17 @@ A few resources to get you started if this is your first brick template:
 - [Official Mason Documentation][2]
 - [Code generation with Mason Blog][3]
 - [Very Good Livestream: Felix Angelov Demos Mason][4]
+- [Flutter Package of the Week: Mason][5]
+- [Observable Flutter: Building a Mason brick][6]
+- [Meet Mason: Flutter Vikings 2022][7]
 
 [1]: https://github.com/felangel/mason
-[2]: https://github.com/felangel/mason/tree/master/packages/mason_cli#readme
+[2]: https://docs.brickhub.dev
 [3]: https://verygood.ventures/blog/code-generation-with-mason
 [4]: https://youtu.be/G4PTjA6tpTU
+[5]: https://youtu.be/qjA0JFiPMnQ
+[6]: https://youtu.be/o8B1EfcUisw
+[7]: https://youtu.be/LXhgiF5HiQg
 ''';
 
   static const _brickChangelogContent = '''