+++ packages/mason_cli/lib/src/install_brick.dart
@@ -54,46 +54,71 @@ mixin InstallBrickMixin on MasonCommand {
     final resolvedBricks = <String, BrickLocation>{};
     final message = upgrade ? 'Upgrading bricks' : 'Getting bricks';
     final getBricksProgress = logger.progress(message);
+
+    Future<CachedBrick> _resolveBrickEntry(
+      MapEntry<String, BrickLocation> entry,
+    ) async {
+      return () async {
+        final location = resolveBrickLocation(
+          location: entry.value,
+          lockedLocation: upgrade ? null : lockJson.bricks[entry.key],
+        );
+        final masonYamlFile = global ? globalMasonYamlFile : localMasonYamlFile;
+        final normalizedLocation = location.path != null
+            ? BrickLocation(
+                path: canonicalize(
+                  p.join(masonYamlFile.parent.path, location.path),
+                ),
+              )
+            : location;
+        getBricksProgress.update('Installing ${entry.key}');
+        final cachedBrick = await bricksJson.add(
+          Brick(name: entry.key, location: normalizedLocation),
+        );
+        resolvedBricks.addAll(
+          <String, BrickLocation>{entry.key: cachedBrick.brick.location},
+        );
+        getBricksProgress.update('Building ${entry.key}');
+        final generator = await MasonGenerator.fromBrick(
+          Brick.path(cachedBrick.path),
+        );
+        await generator.hooks.compile();
+        return cachedBrick;
+      }();
+    }
+
     try {
-      bricksJson.clear();
+      if (upgrade) bricksJson.clear();
       final masonYaml = global ? globalMasonYaml : localMasonYaml;
       if (masonYaml.bricks.entries.isNotEmpty) {
-        await Future.wait(
-          masonYaml.bricks.entries.map(
-            (entry) {
-              return () async {
-                getBricksProgress.update('Resolving ${entry.key}');
-                final location = resolveBrickLocation(
-                  location: entry.value,
-                  lockedLocation: upgrade ? null : lockJson.bricks[entry.key],
-                );
-                final masonYamlFile =
-                    global ? globalMasonYamlFile : localMasonYamlFile;
-                final normalizedLocation = location.path != null
-                    ? BrickLocation(
-                        path: canonicalize(
-                          p.join(masonYamlFile.parent.path, location.path),
-                        ),
-                      )
-                    : location;
-                getBricksProgress.update('Installing ${entry.key}');
-                final cachedBrick = await bricksJson.add(
-                  Brick(name: entry.key, location: normalizedLocation),
-                );
-                resolvedBricks.addAll(
-                  <String, BrickLocation>{
-                    entry.key: cachedBrick.brick.location
-                  },
-                );
-                getBricksProgress.update('Compiling ${entry.key}');
-                final generator = await MasonGenerator.fromBrick(
-                  Brick.path(cachedBrick.path),
-                );
-                await generator.hooks.compile();
-              }();
-            },
-          ),
-        );
+        final entries = _BrickEntries.fromBricks(masonYaml.bricks);
+
+        Future<void> _resolveGitBrickEntries(
+          MapEntry<String, List<MapEntry<String, BrickLocation>>> entry,
+        ) async {
+          final firstBrick = entry.value.first;
+          final cachedBrick = await _resolveBrickEntry(firstBrick);
+          final commitHash = cachedBrick.brick.location.git!.ref;
+
+          if (entry.value.length > 1) {
+            for (final entry in entry.value.sublist(1)) {
+              final git = entry.value.git!;
+              await _resolveBrickEntry(
+                MapEntry(
+                  entry.key,
+                  BrickLocation(
+                    git: GitPath(git.url, path: git.path, ref: commitHash),
+                  ),
+                ),
+              );
+            }
+          }
+        }
+
+        await Future.wait([
+          ...entries.gitEntries.entries.map(_resolveGitBrickEntries),
+          ...entries.nonGitEntries.map(_resolveBrickEntry),
+        ]);
       }
     } finally {
       getBricksProgress.complete(message);
@@ -143,4 +168,39 @@ extension on Map<String, BrickLocation> {
       entries.toList()..sort((a, b) => a.key.compareTo(b.key)),
     );
   }
+}
+
+class _BrickEntries {
+  factory _BrickEntries.fromBricks(Map<String, BrickLocation> bricks) {
+    final nonGitEntries = <MapEntry<String, BrickLocation>>[];
+    final gitEntries = <String, List<MapEntry<String, BrickLocation>>>{};
+
+    for (final entry in bricks.entries) {
+      if (entry.value.git != null) {
+        final path = entry.value.git!.url.replaceAll(r'\', '/');
+        final url = base64.encode(utf8.encode(path));
+        final key = '${url}_${entry.value.git!.ref ?? ''}';
+        if (gitEntries.containsKey(key)) {
+          gitEntries[key]!.add(entry);
+        } else {
+          gitEntries[key] = [entry];
+        }
+      } else {
+        nonGitEntries.add(entry);
+      }
+    }
+
+    return _BrickEntries._(
+      nonGitEntries: nonGitEntries,
+      gitEntries: gitEntries,
+    );
+  }
+
+  const _BrickEntries._({
+    this.nonGitEntries = const [],
+    this.gitEntries = const {},
+  });
+
+  final List<MapEntry<String, BrickLocation>> nonGitEntries;
+  final Map<String, List<MapEntry<String, BrickLocation>>> gitEntries;
 }