+++ packages/mason/CHANGELOG.md
@@ -1,3 +1,20 @@
+# 0.1.0-dev.46
+
+- fix: `MasonBundle` use `fieldRename: FieldRename.snake`
+
+# 0.1.0-dev.45
+
+- feat: add `publishTo` to `MasonBundle`
+
+# 0.1.0-dev.44
+
+- feat: support optional `publish_to` in `brick.yaml`
+
+# 0.1.0-dev.43
+
+- fix: sort contents of `bricks.json`
+- fix: brick git installation algorithm
+
 # 0.1.0-dev.42
 
 - feat: improve generated file output